There were changes in the LOGITUDE_FORMATTER and the LATITUDE_FORMATTER. Now, the user can directly change the hemispheres (north-south and west-east) ticklabels ('N', 'S', 'W', 'E') from the gridliner instance. Also, new functionalities have been added to the gridliner
