+++ lib/cartopy/mpl/gridliner.py
@@ -20,6 +20,8 @@ from cartopy.crs import Projection, _RectangularProjection
 from cartopy.mpl.ticker import (
     LongitudeLocator, LatitudeLocator,
     LongitudeFormatter, LatitudeFormatter)
+    
+from formatters import Gridline_Base    
 
 degree_locator = mticker.MaxNLocator(nbins=9, steps=[1, 1.5, 1.8, 2, 3, 6, 10])
 classic_locator = mticker.MaxNLocator(nbins=9)
@@ -43,7 +45,6 @@ _POLAR_PROJS = (
 def _fix_lons(lons):
     """
     Fix the given longitudes into the range ``[-180, 180]``.
-
     """
     lons = np.array(lons, copy=False, ndmin=1)
     fixed_lons = ((lons + 180) % 360) - 180
@@ -51,201 +52,53 @@ def _fix_lons(lons):
     fixed_lons[(fixed_lons == -180) & (lons > 0)] *= -1
     return fixed_lons
 
-def _fix_lats(lons):
-    """
-    Fix the given longitudes into the range ``[-180, 180]``.
 
-    """
-    lons = np.array(lons, copy=False, ndmin=1)
-    fixed_lons = ((lons + 90) % 180) - 90
-    # Make the positive 180s positive again.
-    fixed_lons[(fixed_lons == -90) & (lons > 0)] *= -1
-    return fixed_lons
-
-
-class BASE_CLASS_FORMATER():
-    def __init__(self,
-                 north_hemisphere_str=' - N',
-                 south_hemisphere_str=' - S',
-                 west_hemisphere_str=' - O', 
-                 east_hemisphere_str=' - L',
-                 degree_symbol=None):
-        
-        self.west_hemisphere_str = west_hemisphere_str
-        self.east_hemisphere_str = east_hemisphere_str
-        self.north_hemisphere_str = north_hemisphere_str
-        self.south_hemisphere_str = south_hemisphere_str
-        
-        if degree_symbol is None:
-            self._DEGREE_SYMBOL = _DEGREE_SYMBOL
-        else:
-            self._DEGREE_SYMBOL = degree_symbol
-    
-        
-    def _lon_hemisphere(self, longitude):
-        """Return the hemisphere (E, W or '' for 0) for the given longitude."""
-        longitude = _fix_lons(longitude)
-        if longitude > 0:
-            hemisphere = self.east_hemisphere_str
-        elif longitude < 0:
-            hemisphere = self.west_hemisphere_str
-            
-        elif np.isclose(longitude, 0, rtol=1e-7):
-            hemisphere = self.east_hemisphere_str
-        else:
-            hemisphere = ''
-        return hemisphere
-    
-    
-    def _east_west_formatted(self, longitude, num_format='g'):
-        fmt_string = u'{longitude:{num_format}}{degree}{hemisphere}'.format(longitude=abs(longitude), num_format=num_format,
-                                 hemisphere=self._lon_hemisphere(longitude),
-                                 degree=self._DEGREE_SYMBOL)
-        
-        
-        return fmt_string
-    
-    def _lat_hemisphere(self, latitude):
-        """Return the hemisphere (N, S or '' for 0) for the given latitude."""
-        
-        latitude = _fix_lats(latitude) 
-        if latitude > 0:
-            hemisphere = self.north_hemisphere_str
-        elif latitude < 0:
-            hemisphere = self.south_hemisphere_str
-            
-        elif np.isclose(latitude, 0, rtol=1e-7):
-            hemisphere = self.north_hemisphere_str
-            
-        else:
-            hemisphere = ''
-        return hemisphere
-    
-    def _north_south_formatted(self, latitude, num_format='g'):
-        fmt_string = u'{latitude:{num_format}}{degree}{hemisphere}'
-        return fmt_string.format(latitude=abs(latitude), num_format=num_format,
-                                 hemisphere=self._lat_hemisphere(latitude),
-                                 degree=self._DEGREE_SYMBOL)
-    
-
-class LONGITUDE_FORMATTER(BASE_CLASS_FORMATER):
-    
-    def __init__(self, 
-                 west_hemisphere_str=' - O', 
-                 east_hemisphere_str=' - L',
-                 degree_symbol=None):
-    
-        
-        BASE_CLASS_FORMATER.__init__(self,west_hemisphere_str=west_hemisphere_str,
-                                         east_hemisphere_str=east_hemisphere_str,
-                                         degree_symbol=degree_symbol)
-
-     
-        
-        self.XFormatter = mticker.FuncFormatter(lambda v, pos:
-                                           self._east_west_formatted(v)
-                                           )
-        
-    def __call__(self, lon, pos=None):
-        
-        return self.XFormatter(lon, pos)
-    
-    
-    def set_locs(self, line_ticks):
-        self.XFormatter.set_locs(line_ticks)
-
-    
-    
-class LATITUDE_FORMATTER(BASE_CLASS_FORMATER):
-    
-    def __init__(self, 
-                 north_hemisphere_str=' - Norte',
-                 south_hemisphere_str=' - SUl',
-                 degree_symbol=None):
-    
-        
-        BASE_CLASS_FORMATER.__init__(self, north_hemisphere_str=north_hemisphere_str,
-                            south_hemisphere_str=south_hemisphere_str,
-                            degree_symbol=degree_symbol)
-        
-        
-        
-        self.YFormatter = mticker.FuncFormatter(lambda v, pos:
-                                           self._north_south_formatted(v)
-                                           )
-        
-    def __call__(self, lat, lpos=None):
-        
-        return self.YFormatter(lat, lpos)
-    
-    
-    def set_locs(self, line_ticks):
-        self.YFormatter.set_locs(line_ticks)
-
-#########################################################
+def _lon_hemisphere(longitude):
+    """Return the hemisphere (E, W or '' for 0) for the given longitude."""
+    longitude = _fix_lons(longitude)
+    if longitude > 0:
+        hemisphere = 'E'
+    elif longitude < 0:
+        hemisphere = 'W'
+    else:
+        hemisphere = ''
+    return hemisphere
 
 
+def _lat_hemisphere(latitude):
+    """Return the hemisphere (N, S or '' for 0) for the given latitude."""
+    if latitude > 0:
+        hemisphere = 'N'
+    elif latitude < 0:
+        hemisphere = 'S'
+    else:
+        hemisphere = ''
+    return hemisphere
 
 
-
-class _Gridline_Base():
-
-    def __init__(self):
-        pass
-    
-        
-    def set_number_of_ticks(self, nbins=4, locator='xlocator'):
-        '''
-            Description:
-                This is a helper function for setting (inplace) \
-                the maximum number of ticks in the gridliner
-                
-            Parameters:
-                nbins (int): number of bins to use in the given locator (axis)
-                
-                locator(str): the gridliner locator to set the number of bins
-                    Standard value: 'xlocator'
-                    
-            Return
-                self (gridliner instance)
-                
-        '''
-        locator = getattr(self, locator)
-        
-        locator = mticker.MaxNLocator(nbins)
-        
-        return self
-        
-
-    def change_gridline_tick_decimal_separator(self, 
-                                               gridline_tick_formating='{0:.2f}', 
-                                               axis='yaxis', 
-                                               decimal_separator=',', 
-                                               geographical_symbol=_DEGREE_SYMBOL):  
-         
-        def custom_tick_func_formatter(x, y):
-            ticklabel = '{0}'.format(gridline_tick_formating).format(x) 
-            
-            return ticklabel.replace('.', decimal_separator) + geographical_symbol
-         
-        if axis.lower() == 'both':
-     
-            self.xformatter = ticker.FuncFormatter(lambda x, y : custom_tick_func_formatter(x, y))
-            self.yformatter = ticker.FuncFormatter(lambda x, y : custom_tick_func_formatter(x, y))
-                 
-        else:
-            if axis.lower().startswith('x'):
-                self.xformatter = ticker.FuncFormatter(lambda x, y : custom_tick_func_formatter(x, y))
-            else:
-                self.yformatter = ticker.FuncFormatter(lambda x, y : custom_tick_func_formatter(x, y))
-        
-        
-        return self
+def _east_west_formatted(longitude, num_format='g'):
+    fmt_string = '{longitude:{num_format}}{degree}{hemisphere}'
+    return fmt_string.format(longitude=abs(longitude), num_format=num_format,
+                             hemisphere=_lon_hemisphere(longitude),
+                             degree=_DEGREE_SYMBOL)
 
 
-######################################################
+def _north_south_formatted(latitude, num_format='g'):
+    fmt_string = '{latitude:{num_format}}{degree}{hemisphere}'
+    return fmt_string.format(latitude=abs(latitude), num_format=num_format,
+                             hemisphere=_lat_hemisphere(latitude),
+                             degree=_DEGREE_SYMBOL)
 
-class Gridliner(_Gridline_Base):
+
+#: A formatter which turns longitude values into nice longitudes such as 110W
+LONGITUDE_FORMATTER = mticker.FuncFormatter(lambda v, pos:
+                                            _east_west_formatted(v))
+#: A formatter which turns longitude values into nice longitudes such as 45S
+LATITUDE_FORMATTER = mticker.FuncFormatter(lambda v, pos:
+                                           _north_south_formatted(v))
+
+
+class Gridliner(Gridline_Base):
     # NOTE: In future, one of these objects will be add-able to a GeoAxes (and
     # maybe even a plain old mpl axes) and it will call the "_draw_gridliner"
     # method on draw. This will enable automatic gridline resolution
